+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - update deltachat-node and deltachat/jsonrpc-client to `v1.101.0`
   - fixes 100% cpu usage bug
+- Update translations (16.11.2022)
 
 ### Fixed